pyproject.toml
[project]
name = "pdm-classifier-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.12",
    "langchain>=0.3.22",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.50",
    "langchain-google-genai>=2.1.2",
    "langchain-groq>=0.3.2",
    "nltk>=3.9.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=4.0.1",
    "torch>=2.6.0",
    "transformers>=4.50.3",
    "uvicorn>=0.34.0",
]
